+++ RedPandaIDE/settingsdialog/projectcompilerwidget.cpp
@@ -50,6 +50,7 @@ void ProjectCompilerWidget::refreshOptions()
     QByteArray execEncoding = mExecCharset;
     if (execEncoding == ENCODING_AUTO_DETECT
             || execEncoding == ENCODING_SYSTEM_DEFAULT
+            || execEncoding == ENCODING_OEM_DEFAULT
             || execEncoding == ENCODING_UTF8) {
         int index =ui->cbEncoding->findData(execEncoding);
         ui->cbEncoding->setCurrentIndex(index);
@@ -117,7 +118,10 @@ void ProjectCompilerWidget::init()
     ui->cbCompilerSet->blockSignals(false);
     ui->cbEncodingDetails->setVisible(false);
     ui->cbEncoding->clear();
-    ui->cbEncoding->addItem(tr("System Default(%1)").arg(QString(pCharsetInfoManager->getDefaultConsoleEncoding())),ENCODING_SYSTEM_DEFAULT);
+    ui->cbEncoding->addItem(tr("System Default(%1)").arg(QString(pCharsetInfoManager->getDefaultSystemEncoding())),ENCODING_SYSTEM_DEFAULT);
+#ifdef Q_OS_WIN
+    ui->cbEncoding->addItem(tr("OEM(%1)").arg(QString(pCharsetInfoManager->getDefaultConsoleEncoding())),ENCODING_OEM_DEFAULT);
+#endif
     ui->cbEncoding->addItem(tr("UTF-8"),ENCODING_UTF8);
     foreach (const QString& langName, pCharsetInfoManager->languageNames()) {
         ui->cbEncoding->addItem(langName,langName);
@@ -187,6 +191,7 @@ void ProjectCompilerWidget::on_cbEncoding_currentTextChanged(const QString &/*ar
     QString userData = ui->cbEncoding->currentData().toString();
     if (userData == ENCODING_AUTO_DETECT
             || userData == ENCODING_SYSTEM_DEFAULT
+            || userData == ENCODING_OEM_DEFAULT
             || userData == ENCODING_UTF8) {
         ui->cbEncodingDetails->setVisible(false);
         ui->cbEncodingDetails->clear();