+++ NEWS.md
@@ -5,7 +5,7 @@ Red Panda C++ Version 3.3
   - enhancement: Auto scale the main ui & choose theme dialog at the first run.
   - fix: crash on statements like "x::a x;"
   - enhancement: support literal operators.
-  - fix: Use oem code page instead of ANSI code page for compiling.
+  - enhancement: Auto detect windows OEM code page in options / compiler / compiler set auto convert encoding option combobox.
   - enhancement: Show oem codpages in compiler setting's encoding combobox.
 
 Red Panda C++ Version 3.2